Moved post filters to separate component
